internal/compact: move compaction output splitting logic
This commit moves the logic around around when to split a compaction or flush's
keys to new output files into a new internal/compact package. This commit is
preparation for future work that aspires to create more well-defined interfaces
surrounding compactions and their data dependencies.
