+++ compaction_test.go
@@ -2906,100 +2906,6 @@ func TestCompactionCheckOrdering(t *testing.T) {
 		})
 }
 
-type mockSplitter struct {
-	shouldSplitVal maybeSplit
-}
-
-func (m *mockSplitter) shouldSplitBefore(key *InternalKey, tw *sstable.Writer) maybeSplit {
-	return m.shouldSplitVal
-}
-
-func (m *mockSplitter) onNewOutput(key []byte) []byte {
-	return nil
-}
-
-func TestCompactionOutputSplitters(t *testing.T) {
-	var main, child0, child1 compactionOutputSplitter
-	var prevUserKey []byte
-	pickSplitter := func(input string) *compactionOutputSplitter {
-		switch input {
-		case "main":
-			return &main
-		case "child0":
-			return &child0
-		case "child1":
-			return &child1
-		default:
-			t.Fatalf("invalid splitter slot: %s", input)
-			return nil
-		}
-	}
-
-	datadriven.RunTest(t, "testdata/compaction_output_splitters",
-		func(t *testing.T, d *datadriven.TestData) string {
-			switch d.Cmd {
-			case "reset":
-				main = nil
-				child0 = nil
-				child1 = nil
-			case "init":
-				if len(d.CmdArgs) < 2 {
-					return "expected at least 2 args"
-				}
-				splitterToInit := pickSplitter(d.CmdArgs[0].Key)
-				switch d.CmdArgs[1].Key {
-				case "array":
-					*splitterToInit = &splitterGroup{
-						cmp:       base.DefaultComparer.Compare,
-						splitters: []compactionOutputSplitter{child0, child1},
-					}
-				case "mock":
-					*splitterToInit = &mockSplitter{}
-				case "userkey":
-					*splitterToInit = &userKeyChangeSplitter{
-						cmp: base.DefaultComparer.Compare,
-						unsafePrevUserKey: func() []byte {
-							return prevUserKey
-						},
-						splitter: child0,
-					}
-				}
-				(*splitterToInit).onNewOutput(nil)
-			case "set-should-split":
-				if len(d.CmdArgs) < 2 {
-					return "expected at least 2 args"
-				}
-				splitterToSet := (*pickSplitter(d.CmdArgs[0].Key)).(*mockSplitter)
-				var val maybeSplit
-				switch d.CmdArgs[1].Key {
-				case "split-now":
-					val = splitNow
-				case "no-split":
-					val = noSplit
-				default:
-					t.Fatalf("unexpected value for should-split: %s", d.CmdArgs[1].Key)
-				}
-				splitterToSet.shouldSplitVal = val
-			case "should-split-before":
-				if len(d.CmdArgs) < 1 {
-					return "expected at least 1 arg"
-				}
-				key := base.ParseInternalKey(d.CmdArgs[0].Key)
-				shouldSplit := main.shouldSplitBefore(&key, nil)
-				if shouldSplit == splitNow {
-					main.onNewOutput(key.UserKey)
-					prevUserKey = nil
-				} else {
-					prevUserKey = key.UserKey
-				}
-				return shouldSplit.String()
-			default:
-				return fmt.Sprintf("unknown command: %s", d.Cmd)
-			}
-			return "ok"
-		})
-}
-
 func TestCompactFlushQueuedMemTableAndFlushMetrics(t *testing.T) {
 	t.Run("", func(t *testing.T) {
 		// Verify that manual compaction forces a flush of a queued memtable.
