+++ internal/compact/testdata/frontiers
