+++ internal/compact/testdata/output_splitters
